Add files via upload
Scholars Edition of the London Baptist Confession.  Rather than just listing the scripture citation references, this verse prints out the full text of the verse as many versions of the Westminster Confession.  However, there is more!  This scholars edition prints out the full text in multiple English, Greek, Latin, and Hebrew versions as well.  My apologies that it does not include the KJV, but you get that with most of the Westminster Confession versions.